requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
attrs==25.3.0
Authlib==1.6.0
bandit==1.8.5
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dill==0.4.0
distro==1.9.0
dodgy==0.2.1
dparse==0.6.4
filelock==3.16.1
flake8==7.3.0
flake8-polyfill==1.0.2
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
mando==0.7.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
mypy==1.16.1
mypy_extensions==1.1.0
narwhals==1.45.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
openai==1.93.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pbr==6.1.1
pep8-naming==0.10.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prospector==1.17.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydocstyle==6.3.0
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.7
pylint-celery==0.3
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
radon==6.0.1
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requirements-detector==1.4.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safety==3.5.2
safety-schemas==0.0.14
seaborn==0.13.2
semver==3.0.4
setoptconf-tmp==0.3.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
stevedore==5.4.1
streamlit==1.46.1
streamlit-ace==0.1.1
streamlit-agraph==0.0.45
streamlit-camera-input-live==0.2.0
streamlit-elements==0.1.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.4.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vulture==2.14
